+++ Assets/Scripts/GameManagers.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d4cbb7cfabaf9e24bb74c4c5ce07ffc2
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
