+++ Assets/Scripts/GameManagers/Goal.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7fd951aaa2d84f640ab4fb222b29dca8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
