+++ Assets/Scripts/GameManagers/Timer.cs
@@ -7,8 +7,9 @@ public class Timer : MonoBehaviour
 {
     [SerializeField] Score score;
     [SerializeField] int maxTimer = 100;
-    [SerializeField] GameObject gameOverText = null;
-    [SerializeField] Text timerText;
+    //[SerializeField] GameObject gameOverText = null;
+    //[SerializeField] Text timerText;
+    [SerializeField] Goal goal;
 
     float timer = 0;
     int time = 0;
@@ -16,7 +17,7 @@ public class Timer : MonoBehaviour
     void Start()
     {
         time = maxTimer;
-        timerText.text = time.ToString();
+        //timerText.text = time.ToString();
     }
 
     private void FixedUpdate()
@@ -24,25 +25,25 @@ public class Timer : MonoBehaviour
         if (time - (int)timer > 0)
         {
             timer += Time.deltaTime;
-            timerText.text = (time - (int)timer).ToString();
+            //timerText.text = (time - (int)timer).ToString();
             score.ShowScore(timer);
         }
         else 
         {
             //ゲームを止める
-            GameOver();
+            goal.GameOver();
         }
 
     }
 
-    public void GameOver()
-    {
-        gameOverText.SetActive(true);
-        Time.timeScale = 0;
-    }
-    public void GameStart()
-    {
-        gameOverText.SetActive(false);
-        Time.timeScale = 1;
-    }
+    //public void GameOver()
+    //{
+    //    gameOverText.SetActive(true);
+    //    Time.timeScale = 0;
+    //}
+    //public void GameStart()
+    //{
+    //    gameOverText.SetActive(false);
+    //    Time.timeScale = 1;
+    //}
 }
